Relocated Field(Header) class into a standalone file
